ignore dataset_full.txt file
the file was originally provided by the processor, and some instances had a "range" for the optimal solution on the largest datasets. It was meant to provide an upper and lower bound for the solution, but the student decided to simply remove it and keep the upper bound, to assure the simplicity of the project.